requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
slowapi
sqlalchemy
pydantic[email]
python-jose[cryptography]
python-dotenv
bcrypt
